Add files via upload
A  scraper to get data from 'https://opentender.eu/" websites. The scrapper file is in the .py format and scrapper  have a single python class which will be called to get the required data. The output is in the csv format. Also the data is visualized using bar graph.
Libraries used:  BeautifulSoup, requests, pandas matplotlib